+++ djangae/db/transaction.py
@@ -13,6 +13,10 @@ from google.appengine.datastore.datastore_rpc import (TransactionalConnection,
                                                       TransactionOptions)
 
 
+class PreventedReadError(ValueError):
+    pass
+
+
 def in_atomic_block():
     # At the moment just a wrapper around App Engine so that
     # users don't have to use two different APIs
@@ -22,6 +26,13 @@ def in_atomic_block():
 def _datastore_get_handler(signal, sender, keys, **kwargs):
     txn = current_transaction()
     if txn:
+        for key in keys:
+            if key in txn._protected_keys:
+                raise PreventedReadError(
+                    "Attempted to read key (%s:%s) inside a transaction "
+                    "where it was marked protected" % (key.kind(), key.id_or_name())
+                )
+
         txn._fetched_keys.update(set(keys))
 
 
@@ -41,8 +52,10 @@ class Transaction(object):
         self._previous_connection = None
         self._fetched_keys = set()
         self._put_keys = set()
+        self._protected_keys = set()
 
     def enter(self):
+        self._protected_keys = set()
         self._fetched_keys = set()
         self._put_keys = set()
         self._enter()
@@ -51,6 +64,7 @@ class Transaction(object):
         self._exit()
         self._fetched_keys = set()
         self._put_keys = set()
+        self._protected_keys = set()
 
     def _enter(self):
         raise NotImplementedError()
@@ -81,6 +95,17 @@ class Transaction(object):
 
         return False
 
+    def prevent_read(self, model, pk, connection=None):
+        if not connection:
+            connection = router.db_for_read(model)
+
+        connection = connections[connection]
+        key = rpc.Key.from_path(
+            model._meta.db_table, pk,
+            namespace=connection.settings_dict.get('NAMESPACE', '')
+        )
+
+        self._protected_keys.add(key)
 
     def has_been_read(self, instance, connection=None):
         return self._check_instance_actions(instance, connection, check_fetched=True, check_put=False)
@@ -88,7 +113,7 @@ class Transaction(object):
     def has_been_written(self, instance, connection=None):
         return self._check_instance_actions(instance, connection, check_fetched=False, check_put=True)
 
-    def refresh_if_unread(self, instance):
+    def refresh_if_unread(self, instance, prevent_further_reads=False):
         """
             Calls instance.refresh_from_db() if the instance hasn't already
             been read this transaction. This helps prevent oddities if you
@@ -121,6 +146,11 @@ class Transaction(object):
             return
 
         instance.refresh_from_db()
+
+        # Enable read-protection if the flag was specified
+        if prevent_further_reads:
+            connection = router.db_for_read(instance.__class__, instance=instance)
+            self.prevent_read(type(instance), instance.pk, connection=connection)
 
     def _commit(self):
         if self._connection: